# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PostStatus = typeof PostStatus[keyof typeof PostStatus];


export const PostStatus = {
  scheduled: 'scheduled',
  sent: 'sent',
  failed: 'failed',
  draft: 'draft',
} as const;

export interface Post {
  id: number;
  content: string;
  platforms: string[];
  scheduledAt: string;
  status: PostStatus;
  /** @nullable */
  imageUrl?: string | null;
  hashtags?: string[];
  createdAt: string;
  /** @nullable */
  sentAt?: string | null;
}

export type PostInputStatus = typeof PostInputStatus[keyof typeof PostInputStatus];


export const PostInputStatus = {
  scheduled: 'scheduled',
  draft: 'draft',
} as const;

export interface PostInput {
  /** @minLength 1 */
  content: string;
  /** @minItems 1 */
  platforms: string[];
  scheduledAt: string;
  imageUrl?: string;
  hashtags?: string[];
  status?: PostInputStatus;
}

export type PostUpdateStatus = typeof PostUpdateStatus[keyof typeof PostUpdateStatus];


export const PostUpdateStatus = {
  scheduled: 'scheduled',
  draft: 'draft',
} as const;

export interface PostUpdate {
  content?: string;
  platforms?: string[];
  scheduledAt?: string;
  /** @nullable */
  imageUrl?: string | null;
  hashtags?: string[];
  status?: PostUpdateStatus;
}

export type DmStatus = typeof DmStatus[keyof typeof DmStatus];


export const DmStatus = {
  scheduled: 'scheduled',
  sent: 'sent',
  failed: 'failed',
  draft: 'draft',
} as const;

export interface Dm {
  id: number;
  message: string;
  recipients: string[];
  platform: string;
  scheduledAt: string;
  status: DmStatus;
  createdAt: string;
  /** @nullable */
  sentAt?: string | null;
}

export type DmInputStatus = typeof DmInputStatus[keyof typeof DmInputStatus];


export const DmInputStatus = {
  scheduled: 'scheduled',
  draft: 'draft',
} as const;

export interface DmInput {
  /** @minLength 1 */
  message: string;
  /** @minItems 1 */
  recipients: string[];
  platform: string;
  scheduledAt: string;
  status?: DmInputStatus;
}

export type DmUpdateStatus = typeof DmUpdateStatus[keyof typeof DmUpdateStatus];


export const DmUpdateStatus = {
  scheduled: 'scheduled',
  draft: 'draft',
} as const;

export interface DmUpdate {
  message?: string;
  recipients?: string[];
  platform?: string;
  scheduledAt?: string;
  status?: DmUpdateStatus;
}

export interface PlatformCount {
  platform: string;
  count: number;
}

export interface DashboardStats {
  totalScheduled: number;
  sentToday: number;
  failedTotal: number;
  draftCount: number;
  dmScheduled: number;
  dmSentToday: number;
  upcomingCount: number;
  platformBreakdown: PlatformCount[];
}

export type CalendarItemType = typeof CalendarItemType[keyof typeof CalendarItemType];


export const CalendarItemType = {
  post: 'post',
  dm: 'dm',
} as const;

export interface CalendarItem {
  id: number;
  type: CalendarItemType;
  title: string;
  scheduledAt: string;
  status: string;
  platforms: string[];
}

export interface Platform {
  id: string;
  name: string;
  icon: string;
  connected: boolean;
  /** @nullable */
  handle: string | null;
  /** @nullable */
  followerCount?: number | null;
}

export type ListPostsParams = {
status?: ListPostsStatus;
platform?: string;
};

export type ListPostsStatus = typeof ListPostsStatus[keyof typeof ListPostsStatus];


export const ListPostsStatus = {
  scheduled: 'scheduled',
  sent: 'sent',
  failed: 'failed',
  draft: 'draft',
} as const;

export type ListDmsParams = {
status?: ListDmsStatus;
};

export type ListDmsStatus = typeof ListDmsStatus[keyof typeof ListDmsStatus];


export const ListDmsStatus = {
  scheduled: 'scheduled',
  sent: 'sent',
  failed: 'failed',
  draft: 'draft',
} as const;

export type GetCalendarParams = {
from?: string;
to?: string;
};