# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformCount } from './platformCount';

export interface DashboardStats {
  totalScheduled: number;
  sentToday: number;
  failedTotal: number;
  draftCount: number;
  dmScheduled: number;
  dmSentToday: number;
  upcomingCount: number;
  platformBreakdown: PlatformCount[];
}