# lib/api-zod/src/generated/types/dmInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DmInputStatus } from './dmInputStatus';

export interface DmInput {
  /** @minLength 1 */
  message: string;
  /** @minItems 1 */
  recipients: string[];
  platform: string;
  scheduledAt: Date;
  status?: DmInputStatus;
}