# lib/api-zod/src/generated/types/dmUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DmUpdateStatus } from './dmUpdateStatus';

export interface DmUpdate {
  message?: string;
  recipients?: string[];
  platform?: string;
  scheduledAt?: Date;
  status?: DmUpdateStatus;
}