# lib/api-zod/src/generated/types/platform.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Platform {
  id: string;
  name: string;
  icon: string;
  connected: boolean;
  /** @nullable */
  handle: string | null;
  /** @nullable */
  followerCount?: number | null;
}