# lib/api-zod/src/generated/types/platformCount.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformCount {
  platform: string;
  count: number;
}