# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PostInputStatus } from './postInputStatus';

export interface PostInput {
  /** @minLength 1 */
  content: string;
  /** @minItems 1 */
  platforms: string[];
  scheduledAt: Date;
  imageUrl?: string;
  hashtags?: string[];
  status?: PostInputStatus;
}