# lib/api-zod/src/generated/types/postUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PostUpdateStatus } from './postUpdateStatus';

export interface PostUpdate {
  content?: string;
  platforms?: string[];
  scheduledAt?: Date;
  /** @nullable */
  imageUrl?: string | null;
  hashtags?: string[];
  status?: PostUpdateStatus;
}